add admin root layout
